+++ game2020/game2020/World/Level1.cs
@@ -25,22 +25,5 @@ namespace Game1
                 { 2,1,1,1,1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,}
             }, 64);
         }
-
-        public void GenerateLevel(int[,] Map, int size)
-        {
-            for (int x = 0; x < Map.GetLength(1); x++)
-            {
-                for (int y = 0; y < Map.GetLength(0); y++)
-                {
-                    int number = Map[y, x];
-
-                    if (number > 0)
-                        CollisionTiles.Add(new CollisionTiles(number, new Rectangle(x * size, y * size, size, size)));
-
-                    width = (x + 1) * size;
-                    height = (y + 1) * size;
-                }
-            }
-        }
     }
 }
