+++ game2020/game2020/World/Levels.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Xna.Framework.Graphics;
+﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -19,6 +20,23 @@ namespace game2020.World
             foreach (CollisionTiles tile in collisionTiles)
             {
                 tile.Draw(spriteBatch);
+            }
+        }
+
+        public void GenerateLevel(int[,] Map, int size)
+        {
+            for (int x = 0; x < Map.GetLength(1); x++)
+            {
+                for (int y = 0; y < Map.GetLength(0); y++)
+                {
+                    int number = Map[y, x];
+
+                    if (number > 0)
+                        CollisionTiles.Add(new CollisionTiles(number, new Rectangle(x * size, y * size, size, size)));
+
+                    width = (x + 1) * size;
+                    height = (y + 1) * size;
+                }
             }
         }
     }
